Final Project/Aruppi/pyproject.toml
[tool.poetry]
name = "aruppi_project"
version = "0.1.0"
description = "This is an application of Japanese Culture, called Aruppi."
authors = ["xsarias <xsariasa@udistrital.edu.co>", "cach <cacelish@udistrital.edu.co"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.30"
pydantic = "^2.7.2"
click = "^8.1.7"
uvicorn = "^0.30.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pylint = "^3.2.2"
black = "^24.4.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
cli_script = "aruppi_project.main:main"
